ref(events-v2): Move events endpoints into separate files
This moves the code for each endpoint into its own file, rather than
attempting to put all code for every events related endpoint into this
one file. Follows the pattern more commonly used in the rest of the codebase.
