+++ src/sentry/api/endpoints/organization_events.py
@@ -2,25 +2,18 @@ from __future__ import absolute_import
 
 import logging
 import six
-
-from datetime import timedelta
 from functools import partial
 from rest_framework.response import Response
 
-from sentry import tagstore
 from sentry.api.bases import OrganizationEventsEndpointBase, OrganizationEventsError, NoProjects
 from sentry.api.helpers.events import get_direct_hit_response
 from sentry.api.paginator import GenericOffsetPaginator
 from sentry.api.serializers import EventSerializer, serialize, SimpleEventSerializer
-from sentry.api.serializers.snuba import SnubaTSResultSerializer
 from sentry.models import SnubaEvent
-from sentry.tagstore.snuba.utils import lookup_tags
-from sentry.utils.dates import parse_stats_period
 from sentry.utils.snuba import (
     raw_query,
-    transform_aliases_and_query,
-    SnubaTSResult,
     SnubaError,
+    transform_aliases_and_query,
 )
 from sentry import features
 from sentry.models.project import Project
@@ -151,134 +144,3 @@ class OrganizationEventsEndpoint(OrganizationEventsEndpointBase):
                     del result['project.id']
 
         return results
-
-
-class OrganizationEventsStatsEndpoint(OrganizationEventsEndpointBase):
-
-    def get(self, request, organization):
-        try:
-            snuba_args = self.get_snuba_query_args(request, organization)
-        except OrganizationEventsError as exc:
-            return Response({'detail': exc.message}, status=400)
-        except NoProjects:
-            return Response({'data': []})
-
-        interval = parse_stats_period(request.GET.get('interval', '1h'))
-        if interval is None:
-            interval = timedelta(hours=1)
-
-        rollup = int(interval.total_seconds())
-
-        y_axis = request.GET.get('yAxis', None)
-        if not y_axis or y_axis == 'event_count':
-            aggregations = [('count()', '', 'count')]
-        elif y_axis == 'user_count':
-            aggregations = [
-                ('uniq', 'tags[sentry:user]', 'count'),
-            ]
-            snuba_args['filter_keys']['tags_key'] = ['sentry:user']
-        else:
-            return Response(
-                {'detail': 'Param yAxis value %s not recognized.' % y_axis}, status=400)
-
-        result = raw_query(
-            aggregations=aggregations,
-            orderby='time',
-            groupby=['time'],
-            rollup=rollup,
-            referrer='api.organization-events-stats',
-            limit=10000,
-            **snuba_args
-        )
-
-        serializer = SnubaTSResultSerializer(organization, None, request.user)
-        return Response(
-            serializer.serialize(
-                SnubaTSResult(result, snuba_args['start'], snuba_args['end'], rollup),
-            ),
-            status=200,
-        )
-
-
-class OrganizationEventsHeatmapEndpoint(OrganizationEventsEndpointBase):
-
-    def get(self, request, organization):
-        try:
-            snuba_args = self.get_snuba_query_args(request, organization)
-        except OrganizationEventsError as exc:
-            return Response({'detail': exc.message}, status=400)
-        except NoProjects:
-            return Response({'detail': 'A valid project must be included.'}, status=400)
-
-        try:
-            keys = self._validate_keys(request)
-            self._validate_project_ids(request, organization, snuba_args)
-        except OrganizationEventsError as error:
-            return Response({'detail': six.text_type(error)}, status=400)
-
-        try:
-            tags = lookup_tags(keys, **snuba_args)
-        except (KeyError, SnubaError) as error:
-            logger.info(
-                'api.organization-events-heatmap',
-                extra={
-                    'organization_id': organization.id,
-                    'user_id': request.user.id,
-                    'keys': keys,
-                    'snuba_args': snuba_args,
-                    'error': six.text_type(error)
-                }
-            )
-            return Response({
-                'detail': 'Invalid query.'
-            }, status=400)
-
-        return Response(serialize(tags, request.user))
-
-    def _validate_keys(self, request):
-        keys = request.GET.getlist('key')
-        if not keys:
-            raise OrganizationEventsError('Tag keys must be specified.')
-
-        for key in keys:
-            if not tagstore.is_valid_key(key):
-                raise OrganizationEventsError('Tag key %s is not valid.' % key)
-
-        return keys
-
-    def _validate_project_ids(self, request, organization, snuba_args):
-        project_ids = snuba_args['filter_keys']['project_id']
-
-        has_global_views = features.has(
-            'organizations:global-views',
-            organization,
-            actor=request.user)
-
-        if not has_global_views and len(project_ids) > 1:
-            raise OrganizationEventsError('You cannot view events from multiple projects.')
-
-        return project_ids
-
-
-class OrganizationEventsMetaEndpoint(OrganizationEventsEndpointBase):
-
-    def get(self, request, organization):
-        try:
-            snuba_args = self.get_snuba_query_args(request, organization)
-        except OrganizationEventsError as exc:
-            return Response({'detail': exc.message}, status=400)
-        except NoProjects:
-            return Response({'count': 0})
-
-        data = raw_query(
-            aggregations=[['count()', '', 'count']],
-            referrer='api.organization-event-meta',
-            turbo=True,
-            **snuba_args
-        )['data'][0]
-
-        return Response({
-            # this needs to be multiplied to account for the `TURBO_SAMPLE_RATE`
-            # in snuba
-            'count': data['count'] * 10,
-        })
