+++ src/sentry/api/endpoints/organization_events_stats.py
@@ -1,0 +1,59 @@
+from __future__ import absolute_import
+
+from datetime import timedelta
+from rest_framework.response import Response
+
+from sentry.api.bases import OrganizationEventsEndpointBase, OrganizationEventsError, NoProjects
+from sentry.api.serializers.snuba import SnubaTSResultSerializer
+from sentry.utils.dates import parse_stats_period
+from sentry.utils.snuba import (
+    raw_query,
+    SnubaTSResult,
+)
+
+
+class OrganizationEventsStatsEndpoint(OrganizationEventsEndpointBase):
+
+    def get(self, request, organization):
+        try:
+            snuba_args = self.get_snuba_query_args(request, organization)
+        except OrganizationEventsError as exc:
+            return Response({'detail': exc.message}, status=400)
+        except NoProjects:
+            return Response({'data': []})
+
+        interval = parse_stats_period(request.GET.get('interval', '1h'))
+        if interval is None:
+            interval = timedelta(hours=1)
+
+        rollup = int(interval.total_seconds())
+
+        y_axis = request.GET.get('yAxis', None)
+        if not y_axis or y_axis == 'event_count':
+            aggregations = [('count()', '', 'count')]
+        elif y_axis == 'user_count':
+            aggregations = [
+                ('uniq', 'tags[sentry:user]', 'count'),
+            ]
+            snuba_args['filter_keys']['tags_key'] = ['sentry:user']
+        else:
+            return Response(
+                {'detail': 'Param yAxis value %s not recognized.' % y_axis}, status=400)
+
+        result = raw_query(
+            aggregations=aggregations,
+            orderby='time',
+            groupby=['time'],
+            rollup=rollup,
+            referrer='api.organization-events-stats',
+            limit=10000,
+            **snuba_args
+        )
+
+        serializer = SnubaTSResultSerializer(organization, None, request.user)
+        return Response(
+            serializer.serialize(
+                SnubaTSResult(result, snuba_args['start'], snuba_args['end'], rollup),
+            ),
+            status=200,
+        )
